Tendermint.SDK.Application

Documentation

data ModuleList (ms :: [Component]) r where

Constructors

NilModules :: ModuleList '[] r

(:+) :: Module name check deliver query es deps r -> ModuleList ms r -> ModuleList (Module name check deliver query es deps ': ms) r infixr 5

data Module (name :: Symbol) (check :: *) (deliver :: *) (query :: *) (es :: EffectRow) (deps :: [Component]) (r :: EffectRow)

Constructors

Module

Fields

moduleTxChecker :: RouteTx check r

moduleTxDeliverer :: RouteTx deliver r

moduleQuerier :: RouteQ query r

moduleEval :: forall s. (Members TxEffs s, Members (DependencyEffs deps) s) => forall a. Sem (es :& s) a -> Sem s a

type Component = EffectRow -> Type

class Eval ms (core :: EffectRow) where

Associated Types

type Effs ms core :: EffectRow

Methods

eval :: proxy core -> ModuleList ms r -> forall a. Sem (Effs ms core) a -> Sem (TxEffs :& BaseAppEffs core) a

type family ModuleEffs (m :: Component) :: EffectRow where ...

Equations

ModuleEffs (Module _ _ _ _ es deps) = es :& (DependencyEffs deps :& (TxEffs :& BaseEffs))

ModuleEffs _ = TypeError (Text "ModuleEffs is a partial function defined only on Component")

data HandlersContext alg ms core

Constructors

HandlersContext

Fields

signatureAlgP :: Proxy alg

modules :: ModuleList ms (Effs ms core)

anteHandler :: AnteHandler (Effs ms core)

compileToCore :: forall a. Sem (BaseAppEffs core) a -> Sem core a

baseAppAnteHandler :: Members AuthEffs r => Member (Error AppError) r => AnteHandler r

createIOApp :: forall r. (forall a. Sem r a -> IO a) -> App (Sem r) -> App IO

makeApp :: forall alg ms core. RecoverableSignatureSchema alg => Message alg ~ Digest SHA256 => ToApplication ms (Effs ms core) => HasTxRouter (ApplicationC ms) (Effs ms core) QueryAndMempool => HasTxRouter (ApplicationC ms) (BaseAppEffs core) QueryAndMempool => HasTxRouter (ApplicationD ms) (Effs ms core) Consensus => HasTxRouter (ApplicationD ms) (BaseAppEffs core) Consensus => HasQueryRouter (ApplicationQ ms) (Effs ms core) => HasQueryRouter (ApplicationQ ms) (BaseAppEffs core) => Eval ms core => HandlersContext alg ms core -> App (Sem core)