Tendermint.SDK.Application.Module

Documentation

data Module (name :: Symbol) (check :: Type) (deliver :: Type) (query :: Type) (es :: EffectRow) (deps :: [Component]) (r :: EffectRow)

Constructors

Module

Fields

moduleTxChecker :: RouteTx check r

moduleTxDeliverer :: RouteTx deliver r

moduleQuerier :: RouteQ query r

moduleEval :: forall s. (Members TxEffs s, Members BaseEffs s, Members (DependencyEffs deps) s) => forall a. Sem (es :& s) a -> Sem s a

type Component = EffectRow -> Type

type family ModuleEffs (m :: Component) :: EffectRow where ...

Equations

ModuleEffs (Module _ _ _ _ es deps) = es :& (DependencyEffs deps :& (TxEffs :& BaseEffs))

ModuleEffs _ = TypeError ('Text "ModuleEffs is a partial function defined only on Component")

data ModuleList (ms :: [Component]) r where

Constructors

NilModules :: ModuleList '[] r

(:+) :: Module name check deliver query es deps r -> ModuleList ms r -> ModuleList (Module name check deliver query es deps ': ms) r infixr 5

data Application check deliver query r s

Constructors

Application

Fields

applicationTxChecker :: RouteTx check r

applicationTxDeliverer :: RouteTx deliver r

applicationQuerier :: RouteQ query s

applicationBeginBlocker :: BeginBlock -> Sem r ()

applicationEndBlocker :: EndBlock -> Sem r EndBlockResult

class ToApplication ms r where

Associated Types

type ApplicationC ms :: Type

type ApplicationD ms :: Type

type ApplicationQ ms :: Type

Methods

toApplication :: ModuleList ms r -> Application (ApplicationC ms) (ApplicationD ms) (ApplicationQ ms) r r

hoistApplication :: HasTxRouter check r 'QueryAndMempool => HasTxRouter deliver r 'Consensus => HasQueryRouter query s => (forall a. Sem r a -> Sem r' a) -> (forall a. Sem s a -> Sem s' a) -> Application check deliver query r s -> Application check deliver query r' s'

class Eval ms (core :: EffectRow) where

Associated Types

type Effs ms core :: EffectRow

Methods

eval :: proxy core -> ModuleList ms r -> forall a. Sem (Effs ms core) a -> Sem (TxEffs :& BaseAppEffs core) a

makeApplication :: Eval ms core => ToApplication ms (Effs ms core) => HasTxRouter (ApplicationC ms) (Effs ms core) 'QueryAndMempool => HasTxRouter (ApplicationD ms) (Effs ms core) 'Consensus => HasQueryRouter (ApplicationQ ms) (Effs ms core) => Proxy core -> AnteHandler (Effs ms core) -> ModuleList ms (Effs ms core) -> (BeginBlock -> Sem (Effs ms core) ()) -> (EndBlock -> Sem (Effs ms core) EndBlockResult) -> Application (ApplicationC ms) (ApplicationD ms) (ApplicationQ ms) (TxEffs :& BaseAppEffs core) (QueryEffs :& BaseAppEffs core)

applyAnteHandler :: HasTxRouter check r 'QueryAndMempool => HasTxRouter deliver r 'Consensus => AnteHandler r -> Application check deliver query r s -> Application check deliver query r s