Tendermint.SDK.BaseApp

BaseApp

type BaseEffs = [Metrics, Logger, Resource, Error AppError]

Concrete row of effects for the BaseApp. Note that because there does | not exist an interpreter for an untagged RawStore, you must scope | these effects before they can be interpreted.

defaultCompileToCore :: forall a. Sem (BaseAppEffs CoreEffs) a -> Sem CoreEffs a

defaultCompileToPureCore :: forall a. Sem (BaseAppEffs PureCoreEffs) a -> Sem PureCoreEffs a

type BaseAppEffs core = StoreEffs :& (BaseEffs :& core)

type family (as :: [a]) :& (bs :: [a]) :: [a] where ... infixr 5

This type family gives a nice syntax for combining multiple lists of effects.

Equations

'[] :& bs = bs

(a ': as) :& bs = a ': (as :& bs)

Core Effects

type CoreEffs = '[Reader LogConfig, Reader (Maybe PrometheusEnv), Reader IAVLVersions, Reader GrpcClient, Embed IO]

CoreEffs is one level below BaseAppEffs, and provides one possible | interpretation for its effects to IO.

data Context

Context is the environment required to run CoreEffs to IO

Constructors

Context

Fields

_contextLogConfig :: LogConfig

_contextPrometheusEnv :: Maybe PrometheusEnv

_contextGrpcClient :: GrpcClient

_contextVersions :: IAVLVersions

contextLogConfig :: Lens' Context LogConfig

contextPrometheusEnv :: Lens' Context (Maybe PrometheusEnv)

contextVersions :: Lens' Context IAVLVersions

makeContext :: InitialLogNamespace -> Maybe MetricsScrapingConfig -> IAVLVersions -> GrpcConfig -> IO Context

runCoreEffs :: Context -> forall a. Sem CoreEffs a -> IO a

The standard interpeter for CoreEffs.

Pure Effects

type PureCoreEffs = '[Reader LogConfig, Reader DBVersions, Reader DB, Error AppError, Embed IO]

CoreEffs is one level below BaseAppEffs, and provides one possible | interpretation for its effects to IO.

data PureContext

Context is the environment required to run CoreEffsPure to IO

Constructors

PureContext

Fields

_pureContextLogConfig :: LogConfig

_pureContextDB :: DB

_pureContextVersions :: DBVersions

pureContextLogConfig :: Lens' PureContext LogConfig

pureContextVersions :: Lens' PureContext DBVersions

pureContextDB :: Lens' PureContext DB

makePureContext :: InitialLogNamespace -> IO PureContext

runPureCoreEffs :: PureContext -> forall a. Sem PureCoreEffs a -> IO (Either AppError a)

The standard interpeter for PureCoreEffs.

Store

data ReadStore m a

Read and Write Effects

data WriteStore m a

class RawKey k where

Keys

Methods

rawKey :: Iso' k ByteString

data StoreKey

Constructors

StoreKey

Fields

skPathFromRoot :: [ByteString]

skKey :: ByteString

class RawKey k => IsKey k ns where

Minimal complete definition

Nothing

Associated Types

type Value k ns :: *

Methods

prefix :: Proxy k -> Proxy ns -> ByteString

prefix :: Proxy k -> Proxy ns -> ByteString

data Store ns

newtype KeyRoot ns

Store

Constructors

KeyRoot ByteString

makeStore :: KeyRoot ns -> Store ns

put :: forall k r ns. IsKey k ns => HasCodec (Value k ns) => Member WriteStore r => Store ns -> k -> Value k ns -> Sem r ()

get :: forall k r ns. IsKey k ns => HasCodec (Value k ns) => Members [ReadStore, Error AppError] r => Store ns -> k -> Sem r (Maybe (Value k ns))

delete :: forall k ns r. IsKey k ns => Member WriteStore r => Store ns -> k -> Sem r ()

Query Routes

data Leaf (a :: *)

data QA (a :: *)

data StoreLeaf a

Errors

data AppError

This type represents a common error response for the query, checkTx, | and deliver tx abci-messages.

Constructors

AppError

Fields

appErrorCode :: Word32

appErrorCodespace :: Text

appErrorMessage :: Text

class IsAppError e where

Allows for custom application error types to be coerced into the standard error resposne.

Methods

makeAppError :: e -> AppError

Events

data Event

Constructors

Event

Fields

eventType :: Text

Type of Event

eventAttributes :: [KVPair]

Event attributes

class ToEvent e where

A class representing a type that can be emitted as an event in the | event logs for the deliverTx response.

Minimal complete definition

Nothing

Methods

toEvent :: e -> Event

toEvent :: (Generic e, GToEvent (Rep e)) => e -> Event

newtype ContextEvent t

Special event wrapper to add contextual event_type info

Constructors

ContextEvent t

emit :: ToEvent e => Member (Output Event) r => e -> Sem r ()

logEvent :: forall e r. (ToJSON e, ToEvent e, Select e) => Member Logger r => e -> Sem r ()

Gas

data GasMeter m a

withGas :: forall r a. MemberWithError GasMeter r => GasAmount -> Sem r a -> Sem r a

Logger

data Logger m a

Effect allowing for console logging.

log :: forall r. MemberWithError Logger r => Severity -> Text -> Sem r ()

data LogSelect

Constructors

All

Some [Text]

addContext :: forall r x a. (MemberWithError Logger r, Select x, ToJSON x) => x -> Sem r a -> Sem r a

data Severity

Constructors

Debug

Info

Warning

Error

Exception

class Select a where

Class for selecting object keys for contextual logging

Minimal complete definition

Nothing

Methods

select :: Verbosity -> a -> LogSelect

select :: Verbosity -> a -> LogSelect

data Verbosity

Constructors

V0

V1

V2

V3

Metrics

data Metrics m a

incCount :: forall r. MemberWithError Metrics r => CountName -> Sem r ()

withTimer :: forall r a. MemberWithError Metrics r => HistogramName -> Sem r a -> Sem r a

data CountName

Constructors

CountName

Fields

countName :: Text

countLabels :: [(Text, Text)]

data HistogramName

Constructors

HistogramName

Fields

histogramName :: Text

histogramLabels :: [(Text, Text)]

histogramBuckets :: [Double]

Transaction

type AnteHandler r = forall msg a. Endo (RoutingTx msg -> Sem r a)

data RoutingTx msg where

Constructors

RoutingTx :: Tx alg msg -> RoutingTx msg

type family RouteTx layout (s :: EffectRow) :: *

data RouteContext

Constructors

CheckTx

DeliverTx

data Return a

data msg :~> a

data TypedMessage msg

type TxEffs = [Output Event, GasMeter, WriteStore, ReadStore, Error AppError]

data EmptyTxServer

Constructors

EmptyTxServer

class DefaultCheckTx api (r :: EffectRow) where

Associated Types

type DefaultCheckTxT api r :: *

Methods

defaultCheckTx :: Proxy api -> Proxy r -> DefaultCheckTxT api r

type family VoidReturn (api :: *) :: * where ...

Equations

VoidReturn (a :<|> b) = VoidReturn a :<|> VoidReturn b

VoidReturn (path :> a) = path :> VoidReturn a

VoidReturn (TypedMessage msg :~> Return a) = TypedMessage msg :~> Return ()

Query

type QueryEffs = '[ReadStore, Error AppError]

class QueryData a where

This class is used to parse the 'data' field of the query request message. | The default method assumes that the 'data' is simply the key for the | value being queried.

Minimal complete definition

Nothing

Methods

fromQueryData :: Base64String -> Either String a

toQueryData :: a -> Base64String

fromQueryData :: RawKey a => Base64String -> Either String a

toQueryData :: RawKey a => a -> Base64String

type family RouteQ layout r :: *

A routeQ handler.

data QueryResult a

Constructors

QueryResult

Fields

queryResultData :: a

queryResultIndex :: Int64

queryResultKey :: Base64String

queryResultProof :: Maybe Proof

queryResultHeight :: Int64

storeQueryHandler :: StoreQueryHandler ns h => ns -> h

data EmptyQueryServer

Constructors

EmptyQueryServer

data RouterError

Constructors

ResourceNotFound