Tendermint.SDK.BaseApp.Errors

Documentation

data AppError

This type represents a common error response for the query, checkTx, | and deliver tx abci-messages.

Constructors

AppError

Fields

appErrorCode :: Word32

appErrorCodespace :: Text

appErrorMessage :: Text

class IsAppError e where

Allows for custom application error types to be coerced into the standard error resposne.

Methods

makeAppError :: e -> AppError

queryAppError :: Lens' Query AppError

This lens is used to set the AppError data into the appropriate | response fields for the query abci-message.

txResultAppError :: Lens' TxResult AppError

This lens is used to set the AppError data into the appropriate | response fields for the checkTx/deliverTx abci-message.

data SDKError

These errors originate from the SDK itself. The "sdk" namespace is reserved | for this error type and should not be used in modules or applications.

Constructors

InternalError Text
Something went wrong and we have no idea what.

ParseError Text
Parsing errors for SDK specific types, e.g. RawTransaction or Msg, etc.

UnmatchedRoute Text
The name of the route that failed to match.

OutOfGasException

MessageValidation [Text]

SignatureRecoveryError Text

NonceException Word64 Word64

StoreError Text

GrpcError Text

UnknownAccountError Address

throwSDKError :: Member (Error AppError) r => SDKError -> Sem r a

As of right now it's not expected that one can recover from an SDKError, | so we are throwing them as AppErrors directly.