Tendermint.SDK.BaseApp.Query

Documentation

serveQueryApplication :: HasQueryRouter layout r => Proxy layout -> Proxy r -> RouteQ layout r -> QueryApplication (Sem r)

class HasQueryRouter layout r where

This class is used to construct a router given a layout type. The layout | is constructed using the combinators that appear in the instances here, no other | Servant combinators are recognized.

Associated Types

type RouteQ layout r :: *

A routeQ handler.

Methods

routeQ :: Proxy layout -> Proxy r -> Delayed (Sem r) env QueryRequest (RouteQ layout r) -> Router env r QueryRequest Query

Transform a routeQ handler into a Router.

class StoreQueryHandlers (kvs :: [*]) (ns :: Symbol) r where

Associated Types

type QueryApi kvs :: *

Methods

storeQueryHandlers :: Proxy kvs -> StoreKey ns -> Proxy r -> RouteQ (QueryApi kvs) r

module Tendermint.SDK.BaseApp.Query.Types

emptyQueryServer :: RouteQ EmptyQueryServer r