Tendermint.SDK.BaseApp.Query.Router

Documentation

class HasQueryRouter layout r where

This class is used to construct a router given a layout type. The layout | is constructed using the combinators that appear in the instances here, no other | Servant combinators are recognized.

Associated Types

type RouteQ layout r :: *

A routeQ handler.

Methods

routeQ :: Proxy layout -> Proxy r -> Delayed (Sem r) env QueryRequest (RouteQ layout (QueryEffs :& r)) -> Router env r QueryRequest Query

Transform a routeQ handler into a Router.

hoistQueryRouter :: Proxy layout -> Proxy r -> (forall a. Sem s a -> Sem s' a) -> RouteQ layout s -> RouteQ layout s'

methodRouter :: HasCodec a => Member (Tagged QueryAndMempool ReadStore) r => Delayed (Sem r) env req (Sem (QueryEffs :& r) (QueryResult a)) -> Router env r req Query