Tendermint.SDK.BaseApp.Query.Types

Documentation

data Leaf (a :: *)

data QA (a :: *)

type QueryApplication m = Query -> m Query

data QueryRequest

Constructors

QueryRequest

Fields

queryRequestPath :: Text

queryRequestParamString :: Text

queryRequestData :: Base64String

queryRequestProve :: Bool

queryRequestHeight :: Int64

parseQueryRequest :: Query -> QueryRequest

data QueryArgs a

Constructors

QueryArgs

Fields

queryArgsProve :: Bool

queryArgsData :: a

queryArgsHeight :: Int64

defaultQueryArgs :: QueryArgs ()

data QueryResult a

Constructors

QueryResult

Fields

queryResultData :: a

queryResultIndex :: Int64

queryResultKey :: Base64String

queryResultProof :: Maybe Proof

queryResultHeight :: Int64

class HasCodec a => Queryable a

class representing objects which can be queried via the hs-abci query message. | Here the Name is the leaf of the query url, e.g. if you can access a token | balance of type Balance at "token/balance", then 'Name Balance ~ "balance"'.

Associated Types

type Name a :: Symbol

class FromQueryData a where

This class is used to parse the 'data' field of the query request message. | The default method assumes that the 'data' is simply the key for the | value being queried.

Minimal complete definition

Nothing

Methods

fromQueryData :: Base64String -> Either String a

fromQueryData :: RawKey a => Base64String -> Either String a

data EmptyQueryServer

Constructors

EmptyQueryServer