Tendermint.SDK.BaseApp.Store.Array

Documentation

data Array (a :: Type)

A 'Array a' is an appendable list whose elements can be accessed | by their index. You can also delete from the list, in which case accessing | that index will result in a Nothing.

makeArray :: IsKey k ns => Value k ns ~ Array a => k -> Store ns -> Value k ns

Smart constuctor to make sure we're making a Array from | the appropriate key type.

makeFullStoreKey :: Array a -> Word64 -> StoreKey

append :: Members [Error AppError, ReadStore, WriteStore] r => HasCodec a => a -> Array a -> Sem r ()

Add an item to the end of the list.

length :: Members [Error AppError, ReadStore] r => Array a -> Sem r Word64

foldl :: Members [Error AppError, ReadStore] r => HasCodec a => (b -> a -> b) -> b -> Array a -> Sem r b

modifyAtIndex :: Members [Error AppError, ReadStore, WriteStore] r => HasCodec a => Word64 -> (a -> a) -> Array a -> Sem r (Maybe a)

Modify a list at a particular index, return the | updated value if the element was found.

deleteWhen :: Members [Error AppError, ReadStore, WriteStore] r => HasCodec a => (a -> Bool) -> Array a -> Sem r ()

Delete when the predicate evaluates to true.

(!!) :: Members [Error AppError, ReadStore] r => HasCodec a => Array a -> Word64 -> Sem r (Maybe a) infixl 9

Access an item directly by its index.

elemIndex :: Members [Error AppError, ReadStore] r => HasCodec a => Eq a => a -> Array a -> Sem r (Maybe Word64)

Get the first index where an element appears in the list.

toList :: Members [Error AppError, ReadStore] r => HasCodec a => Array a -> Sem r [a]

View the Array as a Array.