Tendermint.SDK.BaseApp.Store.TH

Documentation

makeSubStore

Arguments

:: Name
store

-> String
substoreName

-> TypeQ
store type

-> ByteString
key

-> Q [Dec]

class RawKey k => IsKey k ns where

Minimal complete definition

Nothing

Associated Types

type Value k ns :: *

Methods

prefix :: Proxy k -> Proxy ns -> ByteString

prefix :: Proxy k -> Proxy ns -> ByteString

class RawKey k where

Keys

Methods

rawKey :: Iso' k ByteString

data Var (a :: *)

makeVar :: IsKey k ns => Value k ns ~ Var a => k -> Store ns -> Value k ns

data Array (a :: *)

A 'Array a' is an appendable list whose elements can be accessed | by their index. You can also delete from the list, in which case accessing | that index will result in a Nothing.

makeArray :: IsKey k ns => Value k ns ~ Array a => k -> Store ns -> Value k ns

Smart constuctor to make sure we're making a Array from | the appropriate key type.

data List (a :: *)

makeList :: IsKey key ns => Value key ns ~ List a => key -> Store ns -> Value key ns

data Map (k :: *) (v :: *)

makeMap :: IsKey key ns => Value key ns ~ Map k v => key -> Store ns -> Value key ns