Tendermint.SDK.Crypto

Documentation

class MakeDigest a where

Class encapsulating data which can hashed.

Methods

makeDigest :: a -> Digest SHA256

class SignatureSchema alg where

Defines the types and methods for the signature schema parameterized by alg.

Associated Types

type PubKey alg :: *

type PrivateKey alg :: *

type Signature alg :: *

type Message alg :: *

Methods

algorithm :: Proxy alg -> Text

sign :: Proxy alg -> PrivateKey alg -> Message alg -> Signature alg

verify :: Proxy alg -> PubKey alg -> Signature alg -> Message alg -> Bool

makePubKey :: Proxy alg -> ByteString -> Maybe (PubKey alg)

makeSignature :: Proxy alg -> ByteString -> Maybe (Signature alg)

derivePubKey :: Proxy alg -> PrivateKey alg -> PubKey alg

addressFromPubKey :: Proxy alg -> PubKey alg -> Address

class SignatureSchema alg => RecoverableSignatureSchema alg where

Class allowing for signing and recovering signatures for messages.

Associated Types

type RecoverableSignature alg :: *

Methods

signRecoverableMessage :: Proxy alg -> PrivateKey alg -> Message alg -> RecoverableSignature alg

recover :: Proxy alg -> RecoverableSignature alg -> Message alg -> Maybe (PubKey alg)

serializeRecoverableSignature :: Proxy alg -> RecoverableSignature alg -> ByteString

makeRecoverableSignature :: Proxy alg -> ByteString -> Maybe (RecoverableSignature alg)

parsePubKey :: SignatureSchema alg => Proxy alg -> PubKey -> Either Text (PubKey alg)

data Secp256k1