Tendermint.SDK.Types.Message

Documentation

data Msg msg

The basic message format embedded in any transaction.

Constructors

Msg

Fields

msgAuthor :: Address

msgData :: msg

msgType :: Text

class HasMessageType msg where

Methods

messageType :: Proxy msg -> Text

data TypedMessage

Constructors

TypedMessage

Fields

typedMsgData :: ByteString

typedMsgType :: Text

data MessageParseError

This is a general error type, primarily accomodating protobuf messages being parsed | by either the proto3-wire | or the proto-lens libraries.

Constructors

WireTypeError Text
A WireTypeError occurs when the type of the data in the protobuf binary format does not match the type encountered by the parser.

BinaryError Text
A BinaryError occurs when we can't successfully parse the contents of the field.

EmbeddedError Text (Maybe MessageParseError)
An EmbeddedError occurs when we encounter an error while parsing an embedded message.

OtherParseError Text
Unknown or unstructured parsing error.

formatMessageParseError :: MessageParseError -> Text

Useful for returning in error logs or console logging.

coerceProto3Error :: ParseError -> MessageParseError

coerceProtoLensError :: String -> MessageParseError

data MessageSemanticError

Used during message validation to indicate that although the message has parsed | correctly, it fails certain sanity checks.

Constructors

PermissionError Text
Used to indicate that the message signer does not have the authority to send | this message.

InvalidFieldError Text
Used to indicate that a field isn't valid, e.g. enforces non-negative quantities | or nonempty lists.

OtherSemanticError Text

formatMessageSemanticError :: MessageSemanticError -> Text

class ValidateMessage msg where

Methods

validateMessage :: Msg msg -> Validation [MessageSemanticError] ()

nonEmptyCheck :: Eq a => Monoid a => Text -> a -> Validation [MessageSemanticError] ()

isAuthorCheck :: Text -> Msg msg -> (msg -> Address) -> Validation [MessageSemanticError] ()