Tendermint.SDK.BaseApp

BaseApp

type BaseAppEffs = [RawStore, Metrics, Logger, Resource, Error AppError]

Concrete row of effects for the BaseApp. Note that because there does | not exist an interpreter for an untagged RawStore, you must scope | these effects before they can be interpreted.

type family (as :: [a]) :& (bs :: [a]) :: [a] where ... infixr 5

This type family gives a nice syntax for combining multiple lists of effects.

Equations

'[] :& bs = bs

(a ': as) :& bs = a ': (as :& bs)

type BaseApp r = BaseAppEffs :& r

type ScopedBaseApp (s :: ConnectionScope) r = ApplyScope s (BaseApp r)

compileToCoreEffs :: AuthTreeGetter s => forall a. Sem (ScopedBaseApp s CoreEffs) a -> Sem CoreEffs a

An intermediary interpeter, bringing BaseApp down to CoreEff.

compileScopedEff :: ScopedEff CoreEffs a -> Sem CoreEffs a

CoreEff

type CoreEffs = '[MergeScopes, Reader LogConfig, Reader (Maybe PrometheusEnv), Reader AuthTreeState, Embed IO]

CoreEffs is one level below BaseAppEffs, and provides one possible | interpretation for its effects to IO.

data Context

Context is the environment required to run CoreEffs to IO

Constructors

Context

Fields

_contextLogConfig :: LogConfig

_contextPrometheusEnv :: Maybe PrometheusEnv

_contextAuthTree :: AuthTreeState

contextLogConfig :: Lens' Context LogConfig

contextPrometheusEnv :: Lens' Context (Maybe PrometheusEnv)

contextAuthTree :: Lens' Context AuthTreeState

makeContext :: InitialLogNamespace -> Maybe MetricsScrapingConfig -> IO Context

runCoreEffs :: Context -> forall a. Sem CoreEffs a -> IO a

The standard interpeter for CoreEffs.

Store

data RawStore m a

class RawKey k where

Methods

rawKey :: Iso' k ByteString

class RawKey k => IsKey k ns where

Minimal complete definition

Nothing

Associated Types

type Value k ns = a | a -> ns k

Methods

prefixWith :: Proxy k -> Proxy ns -> ByteString

prefixWith :: Proxy k -> Proxy ns -> ByteString

newtype StoreKey n

Constructors

StoreKey ByteString

put :: forall k r ns. IsKey k ns => HasCodec (Value k ns) => Member RawStore r => StoreKey ns -> k -> Value k ns -> Sem r ()

get :: forall k r ns. IsKey k ns => HasCodec (Value k ns) => Members [RawStore, Error AppError] r => StoreKey ns -> k -> Sem r (Maybe (Value k ns))

delete :: forall k ns r. IsKey k ns => Member RawStore r => StoreKey ns -> k -> Sem r ()

Query Routes

data Leaf (a :: *)

data QA (a :: *)

Scope

data ConnectionScope

Constructors

Query

Mempool

Consensus

applyScope :: forall s r. forall a. Sem (RawStore ': r) a -> Sem (Tagged s RawStore ': r) a

Errors

data AppError

This type represents a common error response for the query, checkTx, | and deliver tx abci-messages.

Constructors

AppError

Fields

appErrorCode :: Word32

appErrorCodespace :: Text

appErrorMessage :: Text

class IsAppError e where

Allows for custom application error types to be coerced into the standard error resposne.

Methods

makeAppError :: e -> AppError

data SDKError

These errors originate from the SDK itself. The "sdk" namespace is reserved | for this error type and should not be used in modules or applications.

Constructors

InternalError
Something went wrong and we have no idea what.

ParseError Text
Parsing errors for SDK specific types, e.g. RawTransaction or Msg, etc.

UnmatchedRoute Text
The name of the route that failed to match.

OutOfGasException

MessageValidation [Text]

SignatureRecoveryError Text

NonceException Word64 Word64

throwSDKError :: Member (Error AppError) r => SDKError -> Sem r a

As of right now it's not expected that one can recover from an SDKError, | so we are throwing them as AppErrors directly.

Events

data Event

Constructors

Event

Fields

eventType :: Text

Type of Event

eventAttributes :: [KVPair]

Event attributes

class ToEvent e where

A class representing a type that can be emitted as an event in the | event logs for the deliverTx response.

Minimal complete definition

makeEventType

Methods

makeEventType :: Proxy e -> String

makeEventData :: e -> [(ByteString, ByteString)]

makeEventData :: ToJSON e => e -> [(ByteString, ByteString)]

newtype ContextEvent t

Special event wrapper to add contextual event_type info

Constructors

ContextEvent t

emit :: ToEvent e => Member (Output Event) r => e -> Sem r ()

logEvent :: forall e r. (ToJSON e, ToEvent e, Select e) => Member Logger r => e -> Sem r ()

Gas

data GasMeter m a

Logger

data Logger m a

Effect allowing for console logging.

log :: forall r. MemberWithError Logger r => Severity -> Text -> Sem r ()

addContext :: forall r x a. (MemberWithError Logger r, Select x, ToJSON x) => x -> Sem r a -> Sem r a

data LogSelect

Constructors

All

Some [Text]

data Severity

Constructors

Debug

Info

Warning

Error

Exception

class Select a where

Class for selecting object keys for contextual logging

Minimal complete definition

Nothing

Methods

select :: Verbosity -> a -> LogSelect

select :: Verbosity -> a -> LogSelect

data Verbosity

Constructors

V0

V1

V2

V3

Metrics

data Metrics m a

incCount :: forall r. MemberWithError Metrics r => CountName -> Sem r ()

withTimer :: forall r a. MemberWithError Metrics r => HistogramName -> Sem r a -> Sem r a

data CountName

Constructors

CountName

Fields

countName :: Text

countLabels :: [(Text, Text)]

data HistogramName

Constructors

HistogramName

Fields

histogramName :: Text

histogramLabels :: [(Text, Text)]

histogramBuckets :: [Double]

Transaction

type TransactionApplication m = RoutingTx ByteString -> m TxResult

data RoutingTx msg where

Constructors

RoutingTx :: Tx alg msg -> RoutingTx msg

data RouteContext

Constructors

CheckTx

DeliverTx

type family RouteTx layout r c :: *

type Return = Return' OnCheckUnit

data msg :~> a

data TypedMessage msg

type TxEffs = [Output Event, GasMeter, Error AppError]

data EmptyTxServer

emptyTxServer :: RouteTx EmptyTxServer r c

serveTxApplication :: HasTxRouter layout r c => Proxy layout -> Proxy r -> Proxy (c :: RouteContext) -> RouteTx layout r c -> TransactionApplication (Sem r)

class DefaultCheckTx api (r :: EffectRow) where

Associated Types

type DefaultCheckTxT api r :: *

Methods

defaultCheckTx :: Proxy api -> Proxy r -> DefaultCheckTxT api r

Query

class HasCodec a => Queryable a

class representing objects which can be queried via the hs-abci query message. | Here the Name is the leaf of the query url, e.g. if you can access a token | balance of type Balance at "token/balance", then 'Name Balance ~ "balance"'.

Associated Types

type Name a :: Symbol

class FromQueryData a where

This class is used to parse the 'data' field of the query request message. | The default method assumes that the 'data' is simply the key for the | value being queried.

Minimal complete definition

Nothing

Methods

fromQueryData :: Base64String -> Either String a

fromQueryData :: RawKey a => Base64String -> Either String a

type family QueryApi kvs :: *

type family RouteQ layout r :: *

A routeQ handler.

data QueryResult a

Constructors

QueryResult

Fields

queryResultData :: a

queryResultIndex :: Int64

queryResultKey :: Base64String

queryResultProof :: Maybe Proof

queryResultHeight :: Int64

storeQueryHandlers :: StoreQueryHandlers kvs ns r => Proxy kvs -> StoreKey ns -> Proxy r -> RouteQ (QueryApi kvs) r

serveQueryApplication :: HasQueryRouter layout r => Proxy layout -> Proxy r -> RouteQ layout r -> QueryApplication (Sem r)

data EmptyQueryServer

emptyQueryServer :: RouteQ EmptyQueryServer r