Tendermint.SDK.BaseApp.Query

Documentation

serveQueryApplication :: HasQueryRouter layout r => Proxy layout -> Proxy r -> RouteQ layout (QueryEffs :& r) -> QueryApplication (Sem r)

Re-Exports

class HasQueryRouter layout r where

This class is used to construct a router given a layout type. The layout | is constructed using the combinators that appear in the instances here, no other | Servant combinators are recognized.

Associated Types

type RouteQ layout r :: *

A routeQ handler.

Methods

routeQ :: Proxy layout -> Proxy r -> Delayed (Sem r) env QueryRequest (RouteQ layout (QueryEffs :& r)) -> Router env r QueryRequest Query

Transform a routeQ handler into a Router.

hoistQueryRouter :: Proxy layout -> Proxy r -> (forall a. Sem s a -> Sem s' a) -> RouteQ layout s -> RouteQ layout s'

class StoreQueryHandlers (kvs :: [*]) ns r where

Associated Types

type QueryApi kvs :: *

Methods

storeQueryHandlers :: Proxy kvs -> StoreKey ns -> Proxy r -> RouteQ (QueryApi kvs) r

type QueryEffs = '[ReadStore, Error AppError]

module Tendermint.SDK.BaseApp.Query.Types