Tendermint.SDK.BaseApp.Query.Router

Documentation

class HasQueryRouter layout r where

This class is used to construct a router given a layout type. The layout | is constructed using the combinators that appear in the instances here, no other | Servant combinators are recognized.

Associated Types

type RouteQ layout r :: *

A routeQ handler.

Methods

routeQ :: Proxy layout -> Proxy r -> Delayed (Sem r) env QueryRequest (RouteQ layout r) -> Router env r QueryRequest Query

Transform a routeQ handler into a Router.

emptyQueryServer :: RouteQ EmptyQueryServer r

methodRouter :: HasCodec b => Delayed (Sem r) env req (Sem r (QueryResult b)) -> Router env r req Query